feature: MVP implementation
Web boundary with 3 endpoints:

GET /
returns a simple form that takes a url and sends it to the POST
endpoint. Form request payload is encoded as text/plain to avoid having
to handle multipart on the server.

POST /
creates a hash of a received url in the body payload. Then looks up the
hash and url submitted against a list of existing urls in memory. If the
hash and url entry exists in the memory, it is returned immediately in
the format of a href string. Otherwise, the entry is appended to the
file `urls` in the app root, as well as appended to the urls in memory.

GET /:hash
Looks up the hash value against the urls in memory, if found it returns
html content that includes a meta refresh tag which will redirect the
browser to the stored url.